optimise static initiliasation
