+++ packages/github-lens-api/src/app.test.ts
@@ -1,6 +1,5 @@
 /* eslint-disable @typescript-eslint/no-unsafe-member-access -- For body access which is always any */
-import type { Stage } from 'common/config';
-import type { LogLevel } from 'common/log/log';
+import type { Repository } from 'common/github/github';
 import type { Express } from 'express';
 import request from 'supertest';
 import { buildApp } from './app';
@@ -9,14 +8,8 @@ describe('github-lens api lambda', () => {
 	let app: Express;
 
 	beforeEach(() => {
-		const config = {
-			dataKeyPrefix: 'prefix',
-			dataBucketName: 'bucket',
-			region: 'eu-west-1',
-			stage: 'DEV' as Stage,
-			logLevel: 'info' as LogLevel,
-		};
-		app = buildApp(config);
+		const repoData = Promise.resolve<Repository[]>([]);
+		app = buildApp(repoData);
 	});
 
 	describe('GET /healthcheck', () => {
