+++ packages/github-lens-api/src/app.ts
@@ -1,15 +1,12 @@
-import path from 'path';
 import { json as jsonBodyParser } from 'body-parser';
-import { getObject, getS3Client } from 'common/aws/s3';
+import type { Repository } from 'common/github/github';
 import cors from 'cors';
 import express from 'express';
 import type { Express } from 'express';
 import asyncHandler from 'express-async-handler';
-import type { Config } from './config';
 
-export function buildApp(config: Config): Express {
+export function buildApp(repoData: Promise<Repository[]>): Express {
 	const app = express();
-	const s3Client = getS3Client(config.region);
 
 	app.use(jsonBodyParser());
 
@@ -27,14 +24,7 @@ export function buildApp(config: Config): Express {
 	app.get(
 		'/repos',
 		asyncHandler(async (req: express.Request, res: express.Response) => {
-			const repoFileLocation = path.join(config.dataKeyPrefix, 'repos.json');
-			const data = await getObject(
-				s3Client,
-				config.dataBucketName,
-				repoFileLocation,
-			);
-
-			res.status(200).json(data);
+			res.status(200).json(await repoData);
 		}),
 	);
